backend/requirements.txt
altair==4.2.2
anyio==3.6.2
asn1crypto==1.4.0
attrs==22.2.0
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
blis==0.7.9
boto3==1.15.18
botocore==1.18.18
bson==0.5.10
cachetools==5.3.0
catalogue==2.0.8
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
confection==0.0.3
cryptography==3.4.8
cymem==2.0.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
ecdsa==0.18.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fastapi==0.92.0
fastjsonschema==2.16.2
filelock==3.6.0
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
isodate==0.6.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyterlab-pygments==0.2.2
langcodes==3.3.0
markdown-it-py==2.1.0
MarkupSafe==2.1.1
mdurl==0.1.2
mistune==2.0.4
msrest==0.6.19
murmurhash==1.0.9
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.1.0
openpyxl==3.1.0
oscrypto==1.2.1
packaging==22.0
pandas==1.2.5
pandocfilters==1.5.0
passlib==1.7.4
pathy==0.10.1
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
preshed==3.0.8
protobuf==3.20.3
pyarrow==5.0.0
pyasn1==0.4.8
pycparser==2.20
pycryptodomex==3.9.8
pydantic==1.10.2
pydeck==0.8.0
Pygments==2.14.0
PyJWT==1.7.1
Pympler==1.0.1
pyOpenSSL==20.0.1
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==13.3.1
rsa==4.9
s3transfer==0.3.3
scikit-learn==1.1.3
scikit-learn-extra==0.2.0
scipy==1.9.3
semver==2.13.0
six==1.15.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
snowflake-connector-python==2.6.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.4
srsly==2.4.5
starlette==0.25.0
streamlit==1.17.0
thinc==8.1.5
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typer==0.7.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.25.11
uvicorn==0.20.0
validators==0.20.0
virtualenv==20.14.1
wasabi==0.10.1
watchdog==2.2.1
webencodings==0.5.1
zipp==3.11.0